Fixing 1 character degenerative split case - refactoring srt tests
